# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Indonesian Postal Code API Documentation",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
    "/api-key": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/provinsi": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/kabupaten/{provinsiId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "provinsiId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/kecamatan/{kabupaten}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "kabupaten",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/kelurahan/{kecamatan}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "kecamatan",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key to access this service"
      }
    }
  }
}